requirements.txt
bcrypt==5.0.0
boto3<2
discord.py==2.6.4
geoip2==5.2.0
ossapi==5.3.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
sendgrid==6.12.5
sqlalchemy==2.0.45
timeago==1.0.16
rosu-pp-py==3.1.0
pydantic-settings==2.12.0